Optimizing Turret Classes, MapaSinTiledApp,

+++ src/main/java/dad/javafx/immunedefense/turrets/model/BaseTurret.java
@@ -1,8 +1,10 @@
 package dad.javafx.immunedefense.turrets.model;
 
 import javafx.beans.property.BooleanProperty;
+import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.IntegerProperty;
 import javafx.beans.property.SimpleBooleanProperty;
+import javafx.beans.property.SimpleDoubleProperty;
 import javafx.beans.property.SimpleIntegerProperty;
 import javafx.scene.image.ImageView;
 
@@ -21,7 +23,7 @@ public class BaseTurret extends Turret {
 	
 	private BooleanProperty base_exists;
 	
-	private IntegerProperty fireRate;
+	private DoubleProperty fireRate;
 	
 	/**
 	 * Este es el constructor vacío que inicializa  las variables,
@@ -31,7 +33,7 @@ public class BaseTurret extends Turret {
 		super();
 		base_exists = new SimpleBooleanProperty();
 		damage = new SimpleIntegerProperty();
-		fireRate = new SimpleIntegerProperty();
+		fireRate = new SimpleDoubleProperty();
 	}
 	
 	/**
@@ -43,18 +45,18 @@ public class BaseTurret extends Turret {
 	 * @param fireRate Es una variable dirigida a la cantidad de balas las cuales puede disparar una torreta por segundo, 
      * en caso de las torretas tipo daño constante, la misma será -1.
 	 */
-	public BaseTurret(int damage, int fireRate, boolean base_exists) {
+	public BaseTurret(int damage, double fireRate, boolean base_exists) {
 		super(damage,fireRate);
 		this.base_exists = new SimpleBooleanProperty(base_exists);
 		this.damage = new SimpleIntegerProperty(damage);
-		this.fireRate = new SimpleIntegerProperty(fireRate);
+		this.fireRate = new SimpleDoubleProperty(fireRate);
 	}
 	
 	public IntegerProperty getDamage() {
 		return damage;
 	}
 	
-	public IntegerProperty getFireRate() {
+	public DoubleProperty getFireRate() {
 		return fireRate;
 	}
