+++ src/main/java/dad/javafx/immunedefense/turrets/model/Turret.java
@@ -1,6 +1,8 @@
 package dad.javafx.immunedefense.turrets.model;
 
+import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.IntegerProperty;
+import javafx.beans.property.SimpleDoubleProperty;
 import javafx.beans.property.SimpleIntegerProperty;
 
 /**
@@ -15,7 +17,7 @@ public class Turret
     
     private IntegerProperty damage;
     
-    private IntegerProperty fireRate;
+    private DoubleProperty fireRate;
     
     /**
      * Este es el constructor vacío de la clase Turrets, la cual no se piden datos,
@@ -24,7 +26,7 @@ public class Turret
     public Turret() {
     	// Incializar Variables
     	this.damage = new SimpleIntegerProperty();
-    	this.fireRate = new SimpleIntegerProperty();
+    	this.fireRate = new SimpleDoubleProperty();
     }
     
     /**
@@ -35,11 +37,11 @@ public class Turret
      * @param fireRate Es una variable dirigida a la cantidad de balas las cuales puede disparar una torreta por segundo, 
      * en caso de las torretas tipo daño constante, la misma será -1.
      */
-    public Turret(int damage, int fireRate)
+    public Turret(int damage, double fireRate)
     {
     	// Incializar Variables
     	this.damage = new SimpleIntegerProperty();
-    	this.fireRate = new SimpleIntegerProperty();
+    	this.fireRate = new SimpleDoubleProperty();
     	
         //Añadir Valores;
         this.damage.set(damage);
@@ -50,7 +52,7 @@ public class Turret
 		return damage;
 	}
 	
-	public IntegerProperty getFireRate() {
+	public DoubleProperty getFireRate() {
 		return fireRate;
 	}
 }
